add and get competency should work
