+++ src/app.ts
@@ -1,12 +1,13 @@
 import express, { Application } from 'express';
-import { addCompetencyToPerson, getApplicants, getCompetencies, loginPerson, registerPerson } from './controllers/personController';
+import { getApplicants, loginPerson, registerPerson } from './controllers/personController';
+import {getCompetencies, addCompetencyToPerson } from './controllers/competenceController'
 import { validateCompetencyAdd, validateLogin, validateRegistration } from './middleware/validate';
 import * as dotenv from 'dotenv';
 import { validateAdmin, validateAdminOrOwner } from './middleware/validateAdmin';
 
 const cors = require('cors');
 
-
+// getCompetencies  addCompetencyToPerson
 const corsOptions = {
   origin: 'http://localhost:3001', // This should match the domain of your frontend application
   optionsSuccessStatus: 200 // some legacy browsers (IE11, various SmartTVs) choke on 204
@@ -22,7 +23,8 @@ app.post('/register', validateRegistration, registerPerson);
 app.post('/login', validateLogin, loginPerson);
 app.get('/getApplicants', validateAdmin, getApplicants);
 app.post('/addCompetencyToPerson', validateCompetencyAdd, addCompetencyToPerson);
-app.get('/getCompetencies', validateAdminOrOwner,  getCompetencies);
+app.get('/getCompetencies', validateAdminOrOwner,  getCompetencies); // add anoter validate
+
 const PORT = process.env.PORT || 3000;
 app.listen(PORT, () => {
     console.log(`Server running on http://localhost:${PORT}`);
